Drops support for Play 2.5 and before.
