+++ play-zipkin-tracing/play/src/main/scala/jp/co/bizreach/trace/play/ZipkinTraceService.scala
@@ -1,0 +1,22 @@
+package jp.co.bizreach.trace.play
+
+import javax.inject.Inject
+
+import akka.actor.ActorSystem
+import brave.Tracing
+import jp.co.bizreach.trace.{ZipkinTraceConfig, ZipkinTraceServiceLike}
+
+import scala.concurrent.ExecutionContext
+
+/**
+ * Class for Zipkin tracing at Play2.6.
+ *
+ * @param tracing a Play's configuration
+ * @param actorSystem a Play's actor system
+ */
+class ZipkinTraceService @Inject() (
+  val tracing: Tracing,
+  actorSystem: ActorSystem) extends ZipkinTraceServiceLike {
+
+  implicit val executionContext: ExecutionContext = actorSystem.dispatchers.lookup(ZipkinTraceConfig.AkkaName)
+}
