--- play-zipkin-tracing/play24/src/main/scala/jp/co/bizreach/trace/play24/implicits/ZipkinTraceImplicits.scala
@@ -1,23 +1,0 @@
-package jp.co.bizreach.trace.play24.implicits
-
-import jp.co.bizreach.trace.{TraceData, ZipkinTraceServiceLike}
-import play.api.mvc.RequestHeader
-
-trait ZipkinTraceImplicits {
-
-  // for injection
-  val tracer: ZipkinTraceServiceLike
-
-  /**
-   * Creates a trace data including a span from request headers.
-   *
-   * @param req the HTTP request header
-   * @return the trace data
-   */
-  implicit def request2trace(implicit req: RequestHeader): TraceData = {
-    TraceData(
-      span = tracer.toSpan(req.headers)((headers, key) => headers.get(key))
-    )
-  }
-
-}
