--- play-zipkin-tracing/play25/src/main/scala/jp/co/bizreach/trace/play25/ZipkinTraceService.scala
@@ -1,37 +1,0 @@
-package jp.co.bizreach.trace.play25
-
-import javax.inject.Inject
-
-import akka.actor.ActorSystem
-import brave.Tracing
-import brave.sampler.Sampler
-import jp.co.bizreach.trace.{ZipkinTraceConfig, ZipkinTraceServiceLike}
-import play.api.Configuration
-import zipkin2.reporter.AsyncReporter
-import zipkin2.reporter.okhttp3.OkHttpSender
-
-import scala.concurrent.ExecutionContext
-
-/**
- * Class for Zipkin tracing at Play2.5.
- *
- * @param conf a Play's configuration
- * @param actorSystem a Play's actor system
- */
-class ZipkinTraceService @Inject() (
-  conf: Configuration,
-  actorSystem: ActorSystem) extends ZipkinTraceServiceLike {
-
-  implicit val executionContext: ExecutionContext = actorSystem.dispatchers.lookup(ZipkinTraceConfig.AkkaName)
-
-  val tracing = Tracing.newBuilder()
-    .localServiceName(conf.getString(ZipkinTraceConfig.ServiceName) getOrElse "unknown")
-    .spanReporter(AsyncReporter.create(OkHttpSender.create(
-      (conf.getOptional[String](ZipkinTraceConfig.ZipkinBaseUrl) getOrElse "http://localhost:9411") + "/api/v2/spans"
-    )))
-    .sampler(conf.getString(ZipkinTraceConfig.ZipkinSampleRate)
-      .map(s => Sampler.create(s.toFloat)) getOrElse Sampler.ALWAYS_SAMPLE
-    )
-    .build()
-
-}
