+++ play-zipkin-tracing/project/build.properties
@@ -1,1 +1,1 @@
-sbt.version = 0.13.15
+sbt.version=1.1.6
